Move default_parameters to configuration namespace
