+++ test/unit_tests/unittest_laserscan_ros_conversions.cpp
@@ -19,13 +19,13 @@
 #include <gmock/gmock.h>
 
 #include "psen_scan_v2_standalone/api/laserscan.h"
-#include "psen_scan_v2_standalone/default_parameters.h"
+#include "psen_scan_v2_standalone/configuration/default_parameters.h"
 
 #include "psen_scan_v2/laserscan_ros_conversions.h"
 
 using namespace psen_scan_v2;
 using namespace psen_scan_v2_standalone;
-using namespace psen_scan_v2_standalone::constants;
+using namespace psen_scan_v2_standalone::configuration::constants;
 
 MATCHER_P(IsReversed, data_vec, "")
 {
